Deal with the case of completely unmasked data
With newer versions of iris, all cube.data is a masked array,
but if no points are masked, the masked is simply `False`.
While this should not happen for tos, this version of the code
deals better with this case and thus allows for easier debugging
of faulty model data.